+++ ui/app/src/modules/System/Sites/create/BlueprintReview.tsx
@@ -21,8 +21,8 @@ import Grid from '@material-ui/core/Grid';
 import Typography from '@material-ui/core/Typography';
 import IconButton from "@material-ui/core/IconButton";
 import EditIcon from '@material-ui/icons/Edit';
-import { SiteState } from "../models/Site";
-import { Blueprint, Parameter } from "../models/Blueprint";
+import { SiteState } from "../../../../models/Site";
+import { Blueprint, Parameter } from "../../../../models/Blueprint";
 import { defineMessages, useIntl } from "react-intl";
 import { Theme } from "@material-ui/core/styles/createMuiTheme";
 import Visibility from "@material-ui/icons/Visibility";
